maps/handlers: show offset in exe when setting a value
Especially with a PIE (position independent executable) and ASLR,
it would be useful to see the offset within the executable as the
found static memory address is at a different memory location each
execution. So determine the start and the end address of the
executable in memory, calculate the offset and show it additionally
when setting the static memory value.