maps/handlers: show offset in exe when setting a value
We detect the regions belonging to the executable and determine
the start and the end address of the executable. With that
information we can detect static memory. Especially with a PIE
(position independent executable) and ASLR, it would be useful to
see the offset within the executable as the found memory address
is at a different memory location each execution. So show that
offset additionally when setting a static memory value.