handlers: show match region info with 'list' command
It makes things much easier if the 'list' command not only shows
the match info but the related region info as well. The region type
and the load offset of the executable or library allow us to
calculate the address relative to the load offset or region start.
This value shows the address used within the binary and helps
bypassing address space layout randomization (ASLR).

The GUI needs to be changed in following commits to recognize the
new format.